renamed common to lib
