+++ yara/src/lib/completionProvider.ts
