+++ yara/src/lib/configuration.ts
