+++ yara/src/lib/definitionProvider.ts
