+++ yara/src/lib/helpers.ts
