+++ yara/src/lib/hoverProvider.ts
