+++ yara/src/lib/referenceProvider.ts
