+++ yara/src/lib/snippetProvider.ts
